fix: Incorrect internal datatype for MySQL DECIMAL

+++ drivers/mysql/internal/datatype_conversion.go
@@ -20,8 +20,8 @@ var mysqlTypeToDataTypes = map[string]types.DataType{
 	// Floating point types
 	"float":   types.Float32,
 	"real":    types.Float32,
-	"decimal": types.Float32,
-	"numeric": types.Float32,
+	"decimal": types.Float64,
+	"numeric": types.Float64,
 	"double":  types.Float64,
 
 	// String types
